pyproject.toml
[tool.poetry]
name = "DataAcquisition"
version = "0.1.0"
description = ""
authors = ["hefadog2","Spencer","Peter"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.1"
matplotlib = "^3.4.3"
spotipy = "^2.19.0"
python-dotenv = "^0.19.2"
pandas-stubs = "^1.2.0"
sh = "^1.14.2"
PyAutoGUI = "^0.9.53"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
bandit = "^1.7.2"
black = "^21.10b0"
isort = "^5.10.1"
pylint = "^2.11.1"
safety = "^1.10.3"
semgrep = "^0.76.2"
mypy = "^0.931"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pylint.format]
max-line-length = "88"

[tool.pylint.master]
extension-pkg-whitelist = "pydantic"

[tool.pylint.messages_control]
disable = "C0330, C0326"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
